Refactor build tool selection out of BloopInstall
